refactor about config
