+++ config/ch5.yml
@@ -1,0 +1,11 @@
+type: ch5
+
+play_data:
+  max_file_num: 800
+
+play:
+  simulation_num_per_move: 200
+  share_mtcs_info_in_self_play: True
+  reset_mtcs_info_per_game: 1
+  thinking_loop: 10
+  use_newest_next_generation_model: True
