+++ requirements.txt
@@ -28,6 +28,7 @@ Markdown==2.6.9
 MarkupSafe==1.0
 matplotlib==2.1.0
 mistune==0.7.4
+moke-config==0.2.1
 nbconvert==5.3.1
 nbformat==4.4.0
 nose==1.3.7
