+++ src/reversi_zero/config.py
@@ -1,5 +1,7 @@
 import os
 
+from moke_config import ConfigBase
+
 
 def _project_dir():
     d = os.path.dirname
@@ -10,36 +12,26 @@ def _data_dir():
     return os.path.join(_project_dir(), "data")
 
 
-class Config:
-    def __init__(self, config_type="mini"):
+class Config(ConfigBase):
+    def __init__(self):
+        self.type = "default"
         self.opts = Options()
         self.resource = ResourceConfig()
         self.gui = GuiConfig()
         self.nboard = NBoardConfig()
-
-        if config_type == "mini":
-            import reversi_zero.configs.mini as c
-        elif config_type == "normal":
-            import reversi_zero.configs.normal as c
-        elif config_type == "alt":
-            import reversi_zero.configs.alt as c
-        elif config_type == "ch5":
-            import reversi_zero.configs.ch5 as c
-        else:
-            raise RuntimeError(f"unknown config_type: {config_type}")
-        self.model = c.ModelConfig()
-        self.play = c.PlayConfig()
-        self.play_data = c.PlayDataConfig()
-        self.trainer = c.TrainerConfig()
-        self.eval = c.EvaluateConfig()
-        self.type = config_type
+        self.model = ModelConfig()
+        self.play = PlayConfig()
+        self.play_data = PlayDataConfig()
+        self.trainer = TrainerConfig()
+        self.eval = EvaluateConfig()
+        self.play_with_human = PlayWithHumanConfig()
 
 
-class Options:
+class Options(ConfigBase):
     new = False
 
 
-class ResourceConfig:
+class ResourceConfig(ConfigBase):
     def __init__(self):
         self.project_dir = os.environ.get("PROJECT_DIR", _project_dir())
         self.data_dir = os.environ.get("DATA_DIR", _data_dir())
@@ -73,13 +65,13 @@ class ResourceConfig:
                 os.makedirs(d)
 
 
-class GuiConfig:
+class GuiConfig(ConfigBase):
     def __init__(self):
         self.window_size = (400, 440)
         self.window_title = "reversi-alpha-zero"
 
 
-class PlayWithHumanConfig:
+class PlayWithHumanConfig(ConfigBase):
     def __init__(self):
         self.parallel_search_num = 8
         self.noise_eps = 0
@@ -90,7 +82,7 @@ class PlayWithHumanConfig:
     def update_play_config(self, pc):
         """
 
-        :param reversi_zero.configs.normal.PlayConfig pc:
+        :param PlayConfig pc:
         :return:
         """
         pc.noise_eps = self.noise_eps
@@ -100,9 +92,100 @@ class PlayWithHumanConfig:
         pc.use_newest_next_generation_model = self.use_newest_next_generation_model
 
 
-class NBoardConfig:
+class NBoardConfig(ConfigBase):
     def __init__(self):
         self.my_name = "RAZ"
         self.read_stdin_timeout = 0.1
         self.simulation_num_per_depth_about = 20
         self.hint_callback_per_sim = 10
+
+
+class EvaluateConfig(ConfigBase):
+    def __init__(self):
+        self.game_num = 200  # 400
+        self.replace_rate = 0.55
+        self.play_config = PlayConfig()
+        self.play_config.simulation_num_per_move = 400
+        self.play_config.thinking_loop = 1
+        self.play_config.change_tau_turn = 0
+        self.play_config.noise_eps = 0
+        self.play_config.disable_resignation_rate = 0
+        self.evaluate_latest_first = True
+
+
+class PlayDataConfig(ConfigBase):
+    def __init__(self):
+        # Max Training Data Size = nb_game_in_file * max_file_num * 8
+        self.multi_process_num = 16
+        self.nb_game_in_file = 2
+        self.max_file_num = 800
+        self.save_policy_of_tau_1 = True
+        self.enable_ggf_data = True
+        self.nb_game_in_ggf_file = 100
+
+
+class PlayConfig(ConfigBase):
+    def __init__(self):
+        self.simulation_num_per_move = 200
+        self.share_mtcs_info_in_self_play = True
+        self.reset_mtcs_info_per_game = 1
+        self.thinking_loop = 10
+        self.required_visit_to_decide_action = 400
+        self.start_rethinking_turn = 8
+        self.c_puct = 1
+        self.noise_eps = 0.25
+        self.dirichlet_alpha = 0.5
+        self.change_tau_turn = 4
+        self.virtual_loss = 3
+        self.prediction_queue_size = 16
+        self.parallel_search_num = 8
+        self.prediction_worker_sleep_sec  = 0.0001
+        self.wait_for_expanding_sleep_sec = 0.00001
+        self.resign_threshold = -0.9
+        self.allowed_resign_turn = 20
+        self.disable_resignation_rate = 0.1
+        self.false_positive_threshold = 0.05
+        self.resign_threshold_delta = 0.01
+        self.policy_decay_turn = 60  # not used
+        self.policy_decay_power = 3
+
+        # Using a solver is a kind of cheating!
+        self.use_solver_turn = 50
+        self.use_solver_turn_in_simulation = 50
+
+        #
+        self.schedule_of_simulation_num_per_move = [
+            (0, 8),
+            (300, 50),
+            (2000, 200),
+        ]
+
+        # True means evaluating 'AlphaZero' method (disable 'eval' worker).
+        # Please change to False if you want to evaluate 'AlphaGo Zero' method.
+        self.use_newest_next_generation_model = True
+
+
+class TrainerConfig(ConfigBase):
+    def __init__(self):
+        self.wait_after_save_model_ratio = 1  # wait after saving model
+        self.batch_size = 256  # 2048
+        self.min_data_size_to_learn = 100000
+        self.epoch_to_checkpoint = 1
+        self.start_total_steps = 0
+        self.save_model_steps = 200
+        self.use_tensorboard = True
+        self.logging_per_steps = 100
+        self.lr_schedules = [
+            (0, 0.01),
+            (150000, 0.001),
+            (300000, 0.0001),
+        ]
+
+
+class ModelConfig(ConfigBase):
+    def __init__(self):
+        self.cnn_filter_num = 256
+        self.cnn_filter_size = 3
+        self.res_layer_num = 10
+        self.l2_reg = 1e-4
+        self.value_fc_size = 256
