+++ src/reversi_zero/manager.py
@@ -2,6 +2,9 @@ import argparse
 
 from logging import getLogger
 
+import yaml
+from moke_config import create_config
+
 from .lib.logger import setup_logger
 from .config import Config
 
@@ -13,8 +16,9 @@ CMD_LIST = ['self', 'opt', 'eval', 'play_gui', 'nboard']
 def create_parser():
     parser = argparse.ArgumentParser()
     parser.add_argument("cmd", help="what to do", choices=CMD_LIST)
+    parser.add_argument("-c", help="specify config yaml", dest="config_file")
     parser.add_argument("--new", help="run from new best model", action="store_true")
-    parser.add_argument("--type", help="use normal setting", default="normal")
+    parser.add_argument("--type", help="deprecated. Please use -c instead")
     parser.add_argument("--total-step", help="set TrainerConfig.start_total_steps", type=int)
     return parser
 
@@ -30,13 +34,19 @@ def setup(config: Config, args):
 def start():
     parser = create_parser()
     args = parser.parse_args()
-    config_type = args.type
+    if args.type:
+        print("I'm very sorry. --type option was deprecated. Please use -c option instead!")
+        return 1
 
-    config = Config(config_type=config_type)
+    if args.config_file:
+        with open(args.config_file, "rt") as f:
+            config = create_config(Config, yaml.load(f))
+    else:
+        config = create_config(Config)
     setup(config, args)
 
     if args.cmd != "nboard":
-        logger.info(f"config type: {config_type}")
+        logger.info(f"config type: {config.type}")
 
     if args.cmd == "self":
         from .worker import self_play
