+++ src/reversi_zero/play_game/gui.py
@@ -13,7 +13,7 @@ logger = getLogger(__name__)
 
 
 def start(config: Config):
-    PlayWithHumanConfig().update_play_config(config.play)
+    config.play_with_human.update_play_config(config.play)
     reversi_model = PlayWithHuman(config)
     app = wx.App()
     Frame(reversi_model, config.gui).Show()
